+++ tests/testthat/test-pdf_create_jacket.R
@@ -2,13 +2,13 @@ test_that("`pdf_create_jacket()`", {
 	skip_if_not_installed("piecepackr")
 	on.exit(rm_temp_pdfs(), add = TRUE)
 
-	f1 <- pdf_create_jacket()
+	f1 <- pdf_create_4x6_jacket()
 	expect_equal(qpdf::pdf_length(f1), 1L)
 
-	f2 <- pdf_create_jacket(front = "red", back = "blue", spine = "green", inner = "grey")
+	f2 <- pdf_create_4x6_jacket(front = "red", back = "blue", spine = "green", inner = "grey")
 	expect_equal(qpdf::pdf_length(f2), 2L)
 
-	f3 <- pdf_create_jacket(
+	f3 <- pdf_create_poker_jacket(
 		front = nullGrob(),
 		back = nullGrob(),
 		spine = nullGrob(),