+++ NEWS.md
@@ -1,3 +1,17 @@
+pnpmisc v0.2.0 (development)
+============================
+
+New features
+------------
+
+* The following PDF creation functions:
+
+  + `pdf_create_4x6_jacket()` creates 4x6 photo storage box jacket pdf files.
+  + `pdf_create_poker_jacket()` creates poker deck storage box jacket pdf files.
+
+* `pdf_create_jacket()` gains `orientation`, `width`, `height`, `depth` arguments.
+* `pdf_create_jacket_instructions()` gains an `orientation` argument.
+
 pnpmisc v0.1.1
 ==============